+++ src/nuitka_config_files/supported_nuitka_version.js
@@ -3,12 +3,9 @@
  * @Author: erduotong
  * @Date: 2023-12-03 16:25:41
  */
-/**
- * 可用的配置文件支持的Nuitka版本
- */
 export default {
-    latest_key: "1.9.*",//最新版本的key
+    latest_key: "2.1.*",//最新版本的key
     versions: { //version:path
-        "1.9.*": "1.9._",
+        "2.1.*": "2.1._",
     },
 };
